Fixed killer mem leak pointed out by Mario
Muchh happier now. In binary_kmer.c, when you try to get
an array of binary kmers from a fasta, we leaked for every one that contained an N.
Fix is to pull out as soon as you see non-ACGT without even malloc-ing.
Also, in the read_sequence_from_fasta there was a mem leak
where we did not free a seq object if returning NULL.