Partially working pop_graph - See verbose
OK - I have a unit test that loads two people,
in the same population, and tries to print out graphs for the
two populations. Looks like it works - prints
out a graph for one pop and not the other
but I need to check
Also need automated way of checking the printed out supernodes.