Steps towards auto-checking of supernodes in tests. see verbose
./run_pop_tests works when person1.fasta and person2.fasta
each only have one read.
Committing a version that has 2 reads in each and segfaults.
An invalid read in the strcat in db_graph.c, according
to valgrind.

More importantly the code is becoming a mess.
You need to improve the filenaming in
db_graph.c and db_graph_population.c