On verge of getting debug code working...see verbose
Can almost pass back up an array of supernodes to caller
of db_graph_traverse_specific_person_or_pop.
Only thing preventing it is that freeing seq, etc
at the end of that function somehow
wrecks the memory that I was passing back.

once that is fixes, all you need to do is sort the
supernodes (having chosen lesser of each with its
reverse complement) and then can compare with expected