Fixed the "wrong base counting" bug. See verbose
My test was wrong - bases are counted correctly.
The test where two people with sequence that shares an Alu
but differs in some bases before and after the Alu, is mostly working.
what remains is checking that when it says each person has precisely
1 supernode, which is the whole read, that it is in fact correct.
Either I do it manually, or I write some script (which may have bugs)
to check it...