Add more tests of chromosome intersection printing. see verbose.
Tests pass. But code is a bit of a mess and thus getting confused.
Your function to print a supernode only if it intersects >1 chromosome
obviously prints nothing in your Alu test, as in fact there is only one chromosome
to intersect!
However the real mess is that when you count supernodes, you count
supernodes touched, not supernodes printed. Also you
are passing down an int and a long to count the same thing. Granted
one is used if is_for_testing=true, and one if false, but it is a mess.

This needs rationalising.