new implementation for db_graph_get_perfect_path - add functions for node_action and node_condition (more details in following lines)
The functions for printing supernodes, tip clipping, and so on, have two new paramenters with funtions to executed at each node and the condition to be used to skip node.

For example in printing supernodes one might decide to mark all nodes as visited and intiated a new supernode only from "non-visited" nodes. These two actions can be more sophisticated, the user has full control.