Fixed things so one tests breaks. See verbose
I have just fixed things so one of the tests is more honest.
In run_dB_graph_tests, the test_deletion_validation test was to see
if we could take a breakpoint region (such as those given us my Matt hurles)
and by loading the reference, only look in that region within our
graph, and look for supernodes that match the ref at the start and end
but not the middle. This is not a good test, as deletion of a chunk of sequence
creates new kmers at the join, which may or may not be in the reference elsewhere.
Anyway - furthermore the test fails, and it does not pull out the deleted sequence
in the test case where I take chromosome 1 and delete some sequence.
So - there is possibly a bug here, or possibly the problem is that I have
deleted something near a hairpin (I am finding more deleted sequence than
really is deleted). Anyway - the test was passing, but only because there
was no assert to see if we called the correct deletion.

I am going to rewrite this test.