Add code to get read len distribution on loading fasta/q. BUT..
Not yet sure it works. Specifically, I need to modify
the code to be sure it works with very long reads.
It is fine for fastq already, where max_read_len really
is a max read length. But for fasta, which may have MASSIVE
reads (ref chroms), I should take the last kmer window,
and then check if this is a full entry. If not, then the length of that
chun should be added to the first window of the next chunk
we read.