Fixed bugs, mostly in tests but one horrible one in real code. See below
Most notably, I now have a much faster implementation of the code
to compare two paths and decompose them into shared/unique
segments. Previous code was O(mn) (I knew this, but
was in a hurry) and now is O(m+n).
This commit includes a fix to this faster code, and tests for it,
as well s the main tests of genotyping
at a simple site and at a simple HLA example.