Partial fix of longstanding process_calls bug...
Found the place where I was sometimes putting a \n
in VCFs, when the 3prime flank was zero length,
and I tried to steal  base from it.
Fixed that, but still dont have a fix for getting the right base
from the ref genome and inserting it. Letting vcf_align
remove those lines for now so I can give a VCF to Oliver