Many updates
1. Fixed error cleaning!
Took out the max length on supernode that can be removed. This is
important because in plasmodium you can have contamination of a different
genome, which makes LONG contigs with low covg.

2. Removed tip clipping from cortex_var.c.

3. Various fixes and updates to run_calls and process_calls.

At this stage of the game, we are (in run_calls) making a big set of calls,
and then making  the union and genotyping on the least k.