Small changes
There is too much output when running tests,
can't see wood for trees, so set things up so
certain prints can be commented out when running tests.
Would be better if I had set it up so they did not print
when DEBUG=1, todo.

Also fixed a couple of tests which should not run unless there
are enough colours