BugFix to squeeze_mem in run_calls.pl
The bug was: squeeze mem counts the number of kmers
in the pool of all samples at the lowest kmer and lowest
cleaning level. If using many kmers, it appears
this does not always work, as can have more kmers at one
of the higher kmer values. So the fix is
1. Try highest and lowest kmers and take max
2. Check that it works!!