Mid-debugging - not in a good state
Attempts to fix edge case in BC, with abutting ref and normal bubble.
Currently a) do not catch missing Danny call, and
b) now print some extra calls with zero covg on all alleles.
Since this is from the independent workflow, I can check the per sample
call, and this does not have zero covg on both alleles,
so this is an artefact of printing somehow