Revert yet again that upper limit on length of supernode that can be cleaned off.
Previously I was losing calls in S.aureus, and it was either
because I had turned off tip clipping, or because I had
removed the upper limit. I found it was due to tip clipping,
and so reverted *both*. I should have just reverted tip clipping.
So now again, tip clipping is on, and no upper limit on length of supernode
that can be error cleaned